requirements.txt
torchaudio==2.1.2
pyannote.audio==3.1.1
pytorch-lightning==2.1.3
seaborn==0.13.2
PyYAML==6.0.1
python-dotenv==1.0.1
matplotlib==3.7.4
pandas==2.0.3
